requirements.txt
praw==7.8.1
pandas==2.3.1
numpy==2.3.2
scikit-learn==1.7.1
nltk==3.9.1
yfinance==0.2.65
python-dotenv==1.1.1
flask==3.1.1
werkzeug==3.1.3
//matplotlib==3.10.5
//seaborn==0.13.2